changed data

+++ db_handle/host_db.php
@@ -1,28 +1,53 @@
 <?php
-require_once(__DIR__ ."/../pdo/connection.php");
-require_once(__DIR__ ."/generate_data.php");
-require_once(__DIR__. "/../install/vendor/autoload.php");
+error_reporting(E_ALL);
+ini_set('display_errors', 1);
 
-if (isset($_POST['db']) || isset($_POST['table']) || isset($_POST['count']) || isset($_POST['times'])){
-    $db = $_POST['db'];
-    $table = $_POST['table'];
-    $count = $_POST['count'];
-    $times = $_POST['times'];
-
-    if (!dbExists($pdo, $db)) {
-        $sql = "DROP DATABASE $db";
-        $pdo->exec($sql);
-    }
-    $sql = "CREATE DATABASE $db";
-    $pdo->exec($sql);
-    $pdo->exec("USE $db");
-
-    operateTable($pdo, $table, $count, tableExists($pdo, $table));
-    dataInsert($pdo, $table, $count, $times);
-    header('Location:../index.php?table_success');
+require_once(__DIR__ . "/generate_data.php");
+require_once(__DIR__ . "/../install/vendor/autoload.php");
+$success = true;
+if (isset($_SERVER['REQUEST_METHOD']) && $_SERVER['REQUEST_METHOD'] == "POST") {
+    $host = isset($_POST['host']) ? $_POST['host'] : null;
+    $user = isset($_POST['user']) ? $_POST['user'] : null;
+    $pass = isset($_POST['pass']) ? $_POST['pass'] : null;
+    $success = true;
 } else {
-    // header('Location:../index.php?table_failed');
-    echo 'failed';
+    // Handle cases where there is no POST request.
+    // You can display an error message or perform other actions here.
+    echo "No POST request received.";
+    $success = false;
+}
+if ($success) {
+    try {
+        $pdo = new PDO("mysql:host=$host;dbname=mysql", $user, $pass);
+        $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+    } catch (PDOException $e) {
+        echo 'Connection failed: ' . $e->getMessage();
+        die();
+    }
+
+    if (isset($_POST['db']) || isset($_POST['table']) || isset($_POST['count']) || isset($_POST['times'])) {
+        $db = $_POST['db'];
+        $table = $_POST['table'];
+        $count = $_POST['count'];
+        $times = $_POST['times'];
+
+
+        if (!dbExists($pdo, $db)) {
+            $sql = "DROP DATABASE $db";
+            $pdo->exec($sql);
+        }
+        $sql = "CREATE DATABASE $db";
+        $pdo->exec($sql);
+        $pdo->exec("USE $db");
+
+        operateTable($pdo, $table, $count, tableExists($pdo, $table));
+        dataInsert($pdo, $table, $count, $times);
+        // header('Location:../index.php?table_success');
+        echo "<script>alert('Table $table created.')</script>";
+    } else {
+        header('Location:../index.php?table_failed');
+        echo 'failed';
+    }
 }
 
 /* Functions */
@@ -37,4 +62,3 @@ function dbExists($pdo, $db)
         echo "Error: " . $e->getMessage();
     }
 }
-?>
